+++ s2stools/clim.py
@@ -11,7 +11,7 @@ def climatology(
     mean_or_std: str = "mean",
     ndays_clim_filter: int = 7,
     hide_warnings: bool = False,
-    groupby: str = "validtime",
+    groupby: str = "leadtime",
     dim_number_non_exist: bool = False,
 ):
     """
@@ -36,7 +36,7 @@ def climatology(
 
         If "leadtime", create climatology along same leadtime, e.g., different forecasts but always for
         leaditme=+4 days, leadtime=+5 days, ...; If "validtime", create climatology along same day-of-year, e.g.,
-        different forecasts but always for March 03, March 04, March 05
+        different forecasts but always for March 03, March 04, March 05. Defaults to "leadtime".
 
     dim_number_non_exist : bool
         Depracated, because it is now automatically checked whether adding dimension "number" is neceesary.
@@ -52,6 +52,11 @@ def climatology(
     --------
     If reftimes +- time window are not available then anomalies are still computed, but climatology is less robust.
 
+    Warnings
+    --------
+    In latest version, groupby="validtime" is not yet implemented to work with the new xarray version.
+    Therefore, default changed to groupby="leadtime".
+
     Notes
     -----
     There my be use cases where no running mean should be used, e.g., for computing anomalies of
@@ -61,11 +66,13 @@ def climatology(
 
     """
 
+    # "number" is a required dimension, if not existent, add it
     if "number" not in data.dims:
         dim_number_non_exist = True
     if dim_number_non_exist:
-        data = data.expand_dims("number").assign_coords(number=[0])
+        data = data.expand_dims(dim=dict(number=[0]))
 
+    # hc_year=0 required, because it is dropped later on using .drop_sel(hc_year=0)
     if 0 not in data.hc_year:
         print(
             "Data seems to include no realtime forecasts. No problem, just to let you know."
@@ -73,27 +80,10 @@ def climatology(
         hc_year = data.hc_year.values.tolist()
         data = data.reindex(hc_year=hc_year + [0])
 
-    if isinstance(data, xr.DataArray):
-        aggregation = xarray.core.groupby.DataArrayGroupByAggregations
-    elif isinstance(data, xr.Dataset):
-        aggregation = xarray.core.groupby.DatasetGroupByAggregations
-    else:
-        aggregation = None
-        raise TypeError(
-            "aggregation impossible, expected data to be dataset or datarray"
-        )
-
-    if mean_or_std == "mean":
-        aggregation_func = aggregation.mean
-        aggregation_dim = "stacked_reftime_leadtime"
-    elif mean_or_std == "std":
-        aggregation_func = aggregation.std
-        aggregation_dim = ["stacked_reftime_leadtime", "hc_year"]
-    else:
-        raise ValueError(f"mean_or_std must be 'mean' or 'std', not '{mean_or_std}'")
-
+    # Iterate reftimes
+    if len(data.reftime) < 0:
+        raise ValueError("data must have at least one reftime")
     clim_list = []
-
     for reftime in tqdm(data.reftime, desc="iterating reftimes"):
         # collect all reftimes within time window
         reftime = reftime.values.astype("datetime64[D]")
@@ -106,97 +96,43 @@ def climatology(
             .isel(hc_year=0)
             .reftime
         )
-
-        # stacked forecasts from hindcast years
-        clim_stacked = data.sel(
+        data_for_clim_of_this_reftime = data.sel(
             reftime=hc_reftimes_in_window, hc_year=(data.hc_year != 0)
-        ).stack(date=("reftime", "leadtime"))
-        if mean_or_std == "mean":
-            clim_stacked = clim_stacked.mean(["hc_year", "number"])
-        elif mean_or_std == "std":
-            clim_stacked = clim_stacked.sel(number=0)
+        )
 
         # compute climatology
-        try:
-            if groupby == "validtime":
-                warn(
-                    "If you dont't know what you are doing you should better use groupby='leadtime'.",
-                    Warning,
-                )
-                clim = (
-                    (
-                        clim_stacked.assign_coords(
-                            days_since_reftime_init=(
-                                "date",
-                                (
-                                    clim_stacked.reftime.values
-                                    - reftime
-                                    + clim_stacked.leadtime.values
-                                ).astype("timedelta64[D]"),
-                            )
-                        )  # days_since_reftime_init measures timedelta relative to initialization
-                        .unstack()
-                        .groupby("days_since_reftime_init")
-                        .map(
-                            aggregation_func, dim=aggregation_dim
-                        )  # average or standardize all days that have validtime, e.g., reftime + 2D
-                        .rolling(
-                            days_since_reftime_init=ndays_clim_filter,
-                            center=True,
-                            min_periods=1,
-                        )
-                        .mean()  # rolling mean
-                        .sel(
-                            days_since_reftime_init=slice(
-                                np.timedelta64(0, "D"), data.leadtime[-1]
-                            )
-                        )  # cut climatology to length of forecast
-                        .rename(days_since_reftime_init="leadtime")
-                    )
-                    .assign_coords(reftime=reftime.astype("datetime64[ns]"))
-                    .expand_dims("reftime")
-                )
-
-            ### remove
-            # compute climatology 2
-            if groupby == "leadtime":
-                if mean_or_std == "mean":
-                    aggregation_dim = "reftime"
-                elif mean_or_std == "std":
-                    aggregation_dim = ["reftime", "hc_year"]
-
-                clim = (
-                    (
-                        clim_stacked.unstack()
-                        .groupby("leadtime")
-                        .map(
-                            aggregation_func, dim=aggregation_dim
-                        )  # average or standardize all days that have validtime, e.g., reftime + 2D
-                        .rolling(
-                            leadtime=ndays_clim_filter,
-                            center=True,
-                            min_periods=1,
-                        )
-                        .mean()  # rolling mean
-                        .sel(
-                            leadtime=slice(np.timedelta64(0, "D"), data.leadtime[-1])
-                        )  # cut climatology to length of forecast
-                    )
-                    .assign_coords(reftime=reftime.astype("datetime64[ns]"))
-                    .expand_dims("reftime")
-                )
-        except AttributeError as e:
-            # Catch AttributeError and raise a warning with custom text
-            custom_warning_text = (
-                "\n\nThe error regarding the '_obj' attribute is typically raised when using "
-                "flox. Consider using\n\n"
-                "with xr.set_options(use_flox=False):\n\ts2stools.clim.climatology(...)\n"
+        if groupby == "validtime":
+            raise NotImplementedError(
+                "s2stools.clim.climatology was modified from v0.3.7 on, \
+                because it was not compatible with a recent version of xarray."
+                'groupby="validtime" is not yet implemented to work with that new xarray version.'
+                'If you really need groupby="validtime",\
+                    consider using an older version of s2stools and xarray.'
             )
-            warn(custom_warning_text, Warning)
+        elif groupby == "leadtime":
+            # perform mean or std aggregation
+            if mean_or_std == "mean":
+                clim_aggregeated = data_for_clim_of_this_reftime.mean(
+                    ["hc_year", "number"]
+                ).mean("reftime")
+            elif mean_or_std == "std":
+                clim_aggregeated = data_for_clim_of_this_reftime.std(
+                    ["hc_year", "number"]
+                ).mean("reftime")
+            else:
+                raise ValueError(
+                    "mean_or_std must be either 'mean' or 'std', but is {}".format(
+                        mean_or_std
+                    )
+                )
 
-            # Re-raise the original AttributeError
-            raise e
-        ###
+            clim = (
+                clim_aggregeated.rolling(
+                    leadtime=ndays_clim_filter,
+                    center=True,
+                    min_periods=1,
+                ).mean()  # rolling mean
+            ).expand_dims(dim=dict(reftime=[reftime]))
 
         if len(clim.leadtime) != len(data.leadtime):
             if not hide_warnings: